requirements.txt
numpy==1.21
easydict==1.9

pandas==1.1.5

pyquaternion==0.9.9
PyYAML==5.4.1
scipy==1.5.4
Shapely==1.7.1
tdqm==0.0.1
tensorboard==2.4.1
torchmetrics==0.4.1
tqdm==4.61.1
scikit-image
thop
prettytable
nuscenes-devkit
ipdb
tqdm
